refactor: ReservationRepository 인터페이스 추가 및 중복 코드 분리
- ReservationConverter : Jdbc에서 Reservation 필드를 전달 / 반환 코드의 중복 제거
- ReservationRepository : 콘솔과 웹에서 모두 사용하는 레포 함수 정의
- JdbcReservationRepository, JdbcTemplateReservationRepository : ReservationRepository의 구현체
