+++ src/main/java/nextstep/console/JdbcReservationRepository.java
@@ -1,35 +1,36 @@
 package nextstep.console;
 
 import nextstep.model.Reservation;
-import nextstep.model.Theme;
+import nextstep.repository.ReservationConverter;
+import nextstep.repository.ReservationRepository;
 
 import java.sql.*;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.Optional;
 
-public class ReservationDAO {
 
+public class JdbcReservationRepository implements ReservationRepository {
+
+    @Override
     public Reservation save(Reservation reservation) {
         String sql = "INSERT INTO reservation (date, time, name, theme_name, theme_desc, theme_price) VALUES (?, ?, ?, ?, ?, ?);";
 
         try (Connection con = createConnection();
-             PreparedStatement ps = con.prepareStatement(sql, new String[]{"id"})
-        ) {
-            ps.setDate(1, Date.valueOf(reservation.getDate()));
-            ps.setTime(2, Time.valueOf(reservation.getTime()));
-            ps.setString(3, reservation.getName());
-            ps.setString(4, reservation.getTheme().getName());
-            ps.setString(5, reservation.getTheme().getDesc());
-            ps.setInt(6, reservation.getTheme().getPrice());
+             PreparedStatement ps = con.prepareStatement(sql, new String[]{"id"})) {
+            ReservationConverter.set(ps, reservation);
             ps.executeUpdate();
 
-            return new Reservation(getId(ps), reservation.getDate(), reservation.getTime(), reservation.getName(), reservation.getTheme());
+            ResultSet resultSet = ps.getGeneratedKeys();
+            resultSet.next();
+            Long id = resultSet.getLong("id");
+            return new Reservation(id, reservation.getDate(), reservation.getTime(), reservation.getName(), reservation.getTheme());
         } catch (SQLException e) {
             throw new RuntimeException(e);
         }
     }
 
+    @Override
     public Optional<Reservation> findById(Long id) {
         String sql = "SELECT date, time, name, theme_name, theme_desc, theme_price FROM reservation WHERE id = ?";
 
@@ -39,13 +40,8 @@ public class ReservationDAO {
             ResultSet rs = ps.executeQuery();
 
             if (rs.next()) {
-                LocalDate date = rs.getDate("date").toLocalDate();
-                LocalTime time = rs.getTime("time").toLocalTime();
-                String name = rs.getString("name");
-                String themeName = rs.getString("theme_name");
-                String themeDesc = rs.getString("theme_desc");
-                Integer themePrice = rs.getInt("theme_price");
-                return Optional.of(new Reservation(id, date, time, name, new Theme(themeName, themeDesc, themePrice)));
+                Reservation reservation = ReservationConverter.get(rs, id);
+                return Optional.of(reservation);
             }
         } catch (SQLException e) {
             throw new RuntimeException(e);
@@ -53,8 +49,10 @@ public class ReservationDAO {
         return Optional.empty();
     }
 
+    @Override
     public Boolean existsByDateAndTime(LocalDate date, LocalTime time) {
         String sql = "SELECT count(*) as count FROM reservation WHERE date=? AND time=?";
+
         try (Connection con = createConnection();
              PreparedStatement ps = con.prepareStatement(sql)) {
             ps.setDate(1, Date.valueOf(date));
@@ -67,21 +65,17 @@ public class ReservationDAO {
         }
     }
 
+    @Override
     public void deleteById(Long id) {
         String sql = "DELETE FROM reservation WHERE id = ?";
+
         try (Connection con = createConnection();
-             PreparedStatement ps = con.prepareStatement(sql);
-        ) {
+             PreparedStatement ps = con.prepareStatement(sql)) {
             ps.setLong(1, id);
             ps.execute();
         } catch (SQLException e) {
             throw new RuntimeException(e);
         }
-    }
-    private static Long getId(PreparedStatement ps) throws SQLException {
-        ResultSet rs = ps.getGeneratedKeys();
-        rs.next();
-        return rs.getLong("id");
     }
 
     private static Connection createConnection() {
