--- src/main/java/nextstep/web/ReservationRepository.java
@@ -1,88 +1,0 @@
-package nextstep.web;
-
-import nextstep.model.Reservation;
-import nextstep.model.Theme;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.EmptyResultDataAccessException;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.PreparedStatementCreator;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.support.GeneratedKeyHolder;
-import org.springframework.jdbc.support.KeyHolder;
-import org.springframework.stereotype.Repository;
-
-import java.sql.Date;
-import java.sql.PreparedStatement;
-import java.sql.Time;
-import java.time.LocalDate;
-import java.time.LocalTime;
-import java.util.Optional;
-
-@Repository
-public class ReservationRepository {
-
-    private JdbcTemplate jdbcTemplate;
-
-    private final RowMapper<Reservation> rowMapper = (resultSet, rowNum) -> {
-        Long id = resultSet.getLong("id");
-        LocalDate date = resultSet.getDate("date").toLocalDate();
-        LocalTime time = resultSet.getTime("time").toLocalTime();
-        String name = resultSet.getString("name");
-        String themeName = resultSet.getString("theme_name");
-        String themeDesc = resultSet.getString("theme_desc");
-        Integer themePrice = resultSet.getInt("theme_price");
-        return new Reservation(id, date, time, name, new Theme(themeName, themeDesc, themePrice));
-    };
-
-    @Autowired
-    public ReservationRepository(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    public Reservation save(Reservation reservation) {
-        KeyHolder keyHolder = new GeneratedKeyHolder();
-        String sql = "INSERT INTO reservation (date, time, name, theme_name, theme_desc, theme_price) VALUES (?, ?, ?, ?, ?, ?)";
-
-        PreparedStatementCreator creator = con -> {
-            PreparedStatement pstmt = con.prepareStatement(sql, new String[]{"id"});
-            pstmt.setDate(1, Date.valueOf(reservation.getDate()));
-            pstmt.setTime(2, Time.valueOf(reservation.getTime()));
-            pstmt.setString(3, reservation.getName());
-            pstmt.setString(4, reservation.getTheme().getName());
-            pstmt.setString(5, reservation.getTheme().getDesc());
-            pstmt.setInt(6, reservation.getTheme().getPrice());
-            return pstmt;
-        };
-
-        jdbcTemplate.update(creator, keyHolder);
-
-        Long id = keyHolder.getKey().longValue();
-
-        return new Reservation(id, reservation.getDate(), reservation.getTime(), reservation.getName(), reservation.getTheme());
-    }
-
-    public Optional<Reservation> findById(long id) {
-        try {
-            String sql = "SELECT id, date, time, name, theme_name, theme_desc, theme_price FROM reservation WHERE id = ?";
-            return Optional.of(jdbcTemplate.queryForObject(sql, rowMapper, id));
-        } catch (EmptyResultDataAccessException e) {
-            return Optional.empty();
-        }
-    }
-
-    public void deleteById(Long id) {
-        String sql = "DELETE FROM reservation WHERE id = ?";
-        jdbcTemplate.update(sql, id);
-    }
-
-    public void deleteAll() {
-        String sql = "DELETE FROM reservation";
-        jdbcTemplate.update(sql);
-    }
-
-    public boolean existsByDateAndTime(LocalDate date, LocalTime time) {
-        String sql = "SELECT count(*) FROM reservation WHERE date=? AND time=?";
-        Integer count = jdbcTemplate.queryForObject(sql, Integer.class, date, time);
-        return  count > 0;
-    }
-}
